move bulk operations to bulk.py
