+++ edgy/core/db/querysets/queryset.py
@@ -10,7 +10,7 @@ from collections.abc import (
     Sequence,
 )
 from functools import cached_property
-from typing import TYPE_CHECKING, Any, Generic, Literal, cast, overload
+from typing import TYPE_CHECKING, Any, Generic, Literal, cast
 
 import sqlalchemy
 
@@ -1001,283 +1001,6 @@ class QuerySet(BaseQuerySet[EdgyModel, EdgyEmbedTarget], Generic[EdgyModel, Edgy
         queryset = self._clone()
         queryset.embed_parent = embed_parent
         return queryset
-
-    @overload
-    async def bulk_create(
-        self, objs: Iterable[dict[str, Any] | EdgyModel], *, resolve_embed: Literal[True]
-    ) -> list[EdgyEmbedTarget]: ...
-
-    @overload
-    async def bulk_create(
-        self, objs: Iterable[dict[str, Any] | EdgyModel], *, resolve_embed: Literal[False] = False
-    ) -> list[EdgyModel]: ...
-
-    async def bulk_create(
-        self, objs: Iterable[dict[str, Any] | EdgyModel], resolve_embed: bool = False
-    ) -> list[EdgyModel] | list[EdgyEmbedTarget]:
-        """
-        Bulk creates multiple records in a single batch operation.
-
-        This method bypasses model-level save hooks (except for pre/post-save) for efficiency,
-        and returns plain instances which **may** are incomplete.
-
-        Args:
-            objs: An iterable of dictionaries or model instances to be created.
-        Returns:
-            list[EdgyModel]: A list of newly created objects.
-                             Warning: for performance reasons no embedding is applied and returned objects
-                             are maybe incomplete (check `can_load` property please).
-        """
-        return cast(
-            "list[EdgyModel] | list[EdgyEmbedTarget]",
-            [
-                tup[0]
-                for tup in await self._bulk_get_update_or_create(
-                    objs=objs,
-                    unique_fields=set(self.model_class.pknames),
-                    unique_columns=self.model_class.pkcolumns,
-                    update=False,
-                    retrieve=False,
-                    resolve_embed=resolve_embed,
-                )
-            ],
-        )
-
-    bulk_insert = bulk_create
-
-    @overload
-    async def bulk_update(
-        self,
-        objs: Iterable[EdgyModel],
-        *,
-        update_fields: Iterable[str] | None = None,
-        unique_fields: Iterable[str] | None = None,
-        resolve_embed: Literal[True],
-    ) -> list[EdgyEmbedTarget]: ...
-
-    @overload
-    async def bulk_update(
-        self,
-        objs: Iterable[EdgyModel],
-        *,
-        update_fields: Iterable[str] | None = None,
-        unique_fields: Iterable[str] | None = None,
-        resolve_embed: Literal[False] = False,
-    ) -> list[EdgyModel]: ...
-
-    async def bulk_update(
-        self,
-        objs: Iterable[dict[str, Any] | EdgyModel],
-        *,
-        update_fields: Iterable[str] | None = None,
-        unique_fields: Iterable[str] | None = None,
-        fields: Iterable[str] | None = None,
-        resolve_embed: bool = False,
-    ) -> list[EdgyModel] | list[EdgyEmbedTarget]:
-        """
-        Bulk updates records in a table based on the provided list of model instances and fields.
-
-        The primary key of each model instance is used to identify the record to update.
-        This operation is performed within a database transaction.
-
-        Args:
-            objs: A list of existing model instances to update.
-            fields: A list of field names that should be updated across all instances.
-
-        Returns:
-            list[EdgyModel]: A list of updated objects.
-                             Warning: for performance reasons no embedding is applied.
-        """
-        if fields is not None:
-            warnings.warn(
-                "Use `update_fields` instead `fields`.", DeprecationWarning, stacklevel=2
-            )
-        _unique_fields: set[str]
-        _unique_columns: Sequence[str]
-        if unique_fields is None:
-            _unique_fields = set(self.model_class.pknames)
-            _unique_columns = self.model_class.pkcolumns
-        else:
-            _unique_fields = set(unique_fields)
-            if not _unique_fields:
-                raise ValueError("`unique_fields` empty.")
-            _unique_columns = tuple(
-                col
-                for field in _unique_fields
-                for col in self.model_class.meta.field_to_column_names[field]
-            )
-        return cast(
-            "list[EdgyModel] | list[EdgyEmbedTarget]",
-            [
-                tup[0]
-                for tup in await self._bulk_get_update_or_create(
-                    objs=objs,
-                    unique_fields=_unique_fields,
-                    unique_columns=_unique_columns,
-                    update_fields=set(
-                        self.model_class.meta.fields.keys()
-                        if update_fields is None
-                        else update_fields
-                    ),
-                    update=True,
-                    retrieve=False,
-                    resolve_embed=True,
-                )
-            ],
-        )
-
-    @overload
-    async def bulk_update_or_create(
-        self,
-        objs: Iterable[dict[str, Any] | EdgyModel],
-        *,
-        update_fields: Iterable[str] | None = None,
-        unique_fields: Iterable[str] | None = None,
-        resolve_embed: Literal[True],
-    ) -> list[tuple[EdgyEmbedTarget, bool]]: ...
-
-    @overload
-    async def bulk_update_or_create(
-        self,
-        objs: Iterable[dict[str, Any] | EdgyModel],
-        *,
-        update_fields: Iterable[str] | None = None,
-        unique_fields: Iterable[str] | None = None,
-        resolve_embed: Literal[False] = False,
-    ) -> list[tuple[EdgyModel, bool]] | list[tuple[EdgyEmbedTarget, bool]]: ...
-
-    async def bulk_update_or_create(
-        self,
-        objs: Iterable[dict[str, Any] | EdgyModel],
-        *,
-        update_fields: Iterable[str] | None = None,
-        unique_fields: Iterable[str] | None = None,
-        resolve_embed: bool = False,
-    ) -> list[tuple[EdgyModel, bool]] | list[tuple[EdgyEmbedTarget, bool]]:
-        """
-        Bulk updates or creates records in a table.
-
-        If records exist based on unique fields, they are retrieved.
-        Otherwise, new records are created.
-
-        Args:
-            objs (list[Union[dict[str, Any], EdgyModel]]): A list of objects or dictionaries.
-
-        Returns:
-            list[tuple[EdgyModel, bool]]: A list of tuples with retrieved or newly created objects in the same order
-                                          as provided and as second entry if the instance was created.
-                                          Warning: for performance reasons no embedding is applied and returned objects
-                                          are maybe incomplete (check `can_load` property please).
-        """
-        _unique_fields: set[str]
-        _unique_columns: Sequence[str]
-        if unique_fields is None:
-            _unique_fields = set(self.model_class.pknames)
-            _unique_columns = self.model_class.pkcolumns
-        else:
-            _unique_fields = set(unique_fields)
-            if not _unique_fields:
-                raise ValueError("`unique_fields` empty.")
-            _unique_columns = tuple(
-                col
-                for field in _unique_fields
-                for col in self.model_class.meta.field_to_column_names[field]
-            )
-        return await self._bulk_get_update_or_create(
-            objs=objs,
-            unique_fields=_unique_fields,
-            unique_columns=_unique_columns,
-            update_fields=set(self.model_class.meta.fields.keys()).difference(
-                self.model_class.pknames
-            )
-            if update_fields is None
-            else set(update_fields),
-            update=True,
-            retrieve=True,
-            resolve_embed=resolve_embed,
-        )
-
-    @overload
-    async def bulk_get_or_create(
-        self,
-        objs: Iterable[dict[str, Any] | EdgyModel],
-        *,
-        unique_fields: Iterable[str] | None = None,
-        resolve_embed: Literal[True],
-    ) -> list[tuple[EdgyEmbedTarget, bool]]:
-        """
-        Args:
-            resolve_embed (True): Enables embedding.
-
-        Returns:
-            list[tuple[EdgyEmbedTarget, bool]]: A list of tuples with retrieved or newly created objects and if created.
-                                                Warning: All models must be loadable (`can_load` property is true)
-                                                otherwise an error is raised
-        """
-
-    @overload
-    async def bulk_get_or_create(
-        self,
-        objs: Iterable[dict[str, Any] | EdgyModel],
-        *,
-        unique_fields: Iterable[str] | None = None,
-        resolve_embed: Literal[False] = False,
-    ) -> list[tuple[EdgyModel, bool]]:
-        """
-        Returns:
-            list[tuple[EdgyModel, bool]]: A list of tuples with retrieved or newly created objects and if created.
-                                          Warning: for performance reasons no embedding is applied by default and
-                                         the returned objects are maybe incomplete (check `can_load` property)
-        """
-
-    async def bulk_get_or_create(
-        self,
-        objs: Iterable[dict[str, Any] | EdgyModel],
-        *,
-        unique_fields: Iterable[str] | None = None,
-        resolve_embed: bool = False,
-    ) -> list[tuple[EdgyModel, bool]] | list[tuple[EdgyEmbedTarget, bool]]:
-        """
-        Bulk gets or creates records in a table.
-
-        If records exist based on unique fields, they are retrieved.
-        Otherwise, new records are created.
-
-        Args:
-            objs (Iterable[Union[dict[str, Any], EdgyModel]]): A list of objects or dictionaries.
-            unique_fields (Iterable[str] | None): Fields that determine uniqueness.
-                                                  If None, pknames are used. If empty it fails.
-        Returns:
-            list[tuple[EdgyModel, bool]]: A list of tuples with retrieved or newly created objects in the same order
-                                          as provided and as second entry if the instance was created.
-                                          Warning: for performance reasons no embedding is applied and returned objects
-                                          are maybe incomplete (check `can_load` property please).
-        """
-
-        _unique_fields: set[str]
-        _unique_columns: Sequence[str]
-        if unique_fields is None:
-            _unique_fields = set(self.model_class.pknames)
-            _unique_columns = self.model_class.pkcolumns
-        else:
-            _unique_fields = set(unique_fields)
-            if not _unique_fields:
-                raise ValueError("`unique_fields` empty.")
-            _unique_columns = tuple(
-                col
-                for field in _unique_fields
-                for col in self.model_class.meta.field_to_column_names[field]
-            )
-        return await self._bulk_get_update_or_create(
-            objs=objs,
-            unique_fields=_unique_fields,
-            unique_columns=_unique_columns,
-            update=False,
-            retrieve=True,
-            resolve_embed=resolve_embed,
-        )
-
-    bulk_select_or_insert = bulk_get_or_create
 
     async def delete(self, use_models: bool = False) -> int:
         """
